refactor: move test methods
