+++ tests/src/FunctionalJavascript/Integration/ContentLockTest.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Drupal\Tests\thunder\FunctionalJavascript\Integration;
+
+use Drupal\Tests\thunder\FunctionalJavascript\ThunderJavascriptTestBase;
+
+/**
+ * Tests integration with the content_lock module.
+ *
+ * @group Thunder
+ */
+class ContentLockTest extends ThunderJavascriptTestBase {
+
+  /**
+   * Testing the content lock integration.
+   */
+  public function testContentLock(): void {
+
+    $node = $this->loadNodeByUuid('0bd5c257-2231-450f-b4c2-ab156af7b78d');
+    $this->drupalGet($node->toUrl('edit-form'));
+    $this->assertSession()->pageTextContains('This content is now locked against simultaneous editing. This content will remain locked if you navigate away from this page without saving or unlocking it.');
+
+    $driver = $this->getSession()->getDriver();
+    $driver->click('//*[@id="edit-unlock"]');
+
+    $driver->click('//*[@id="edit-submit"]');
+    $this->assertSession()->pageTextContains('Lock broken. Anyone can now edit this content.');
+
+    $this->drupalGet($node->toUrl('edit-form'));
+    $loggedInUser = $this->loggedInUser->label();
+
+    $this->drupalLogout();
+
+    // Login with other user.
+    $this->logWithRole(static::$defaultUserRole);
+
+    $this->drupalGet($node->toUrl('edit-form'));
+    $this->assertSession()->pageTextContains('This content is being edited by the user ' . $loggedInUser . ' and is therefore locked to prevent other users changes.');
+  }
+
+}
